+++ src/main/java/io/github/qudtlib/maven/rdfio/filter/Filters.java
@@ -1,5 +1,7 @@
 package io.github.qudtlib.maven.rdfio.filter;
 
+import java.io.File;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import org.apache.jena.rdf.model.Model;
@@ -10,6 +12,9 @@ import org.apache.maven.plugins.annotations.Parameter;
 public class Filters {
     private final List<Filter> filters = new ArrayList<>();
     private Log log;
+
+    @Parameter(defaultValue = "${project.basedir}", readonly = true)
+    protected File basedir;
 
     @Parameter
     public void setInclude(IncludeFilter includeFilter) {
@@ -24,6 +29,32 @@ public class Filters {
     @Parameter
     public void setSparqlUpdate(String update) {
         filters.add(new SparqlUpdateFilter(update));
+    }
+
+    @Parameter
+    public void setSparqlUpdateFile(String filename) {
+        try {
+            filters.add(new SparqlUpdateFileFilter(filename, basedir));
+        } catch (IOException e) {
+            throw new RuntimeException(
+                    String.format("Error loading sparql update from file %s", filename), e);
+        }
+    }
+
+    @Parameter
+    public void setSparqlConstruct(String construct) {
+        filters.add(new SparqlConstructFilter(construct));
+    }
+
+    @Parameter
+    public void setSparqlConstructFile(String filename) {
+        try {
+            filters.add(new SparqlConstructFileFilter(filename, basedir));
+        } catch (IOException e) {
+            throw new RuntimeException(
+                    String.format("Error loading sparql construct query from file %s", filename),
+                    e);
+        }
     }
 
     public List<Filter> getFilters() {